feat: trlst-376 render notification footer correctly.
Notifications sent by the API ascertain the correct email footer and contact address. However we also need to render the notification preview in Caseworker, which naively gets `NOTIFY_BLOCK_FOOTER` and `TRADE_REMEDIES_EMAIL` from the API. This change (unfortunately) duplicates the API's approach to determining the correct footer and email values for a notif preview. This was the simplest approach which didn't require updating the API client.